+++ backend/src/main/java/com/utsav/arts/dtos/paymentDTO/PaymentResponseDTO.java
@@ -1,5 +1,6 @@
 package com.utsav.arts.dtos.paymentDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class PaymentResponseDTO {
@@ -7,10 +8,10 @@ public class PaymentResponseDTO {
     private int id;
     private int orderId;
     private int userId;
-    private int amount;
+    private BigDecimal amount;
     private String currency;
     private String method;
-    private String status;
+    private String status;       // Will be mapped from enum
     private String transactionId;
     private LocalDateTime createdAt;
 
@@ -25,8 +26,8 @@ public class PaymentResponseDTO {
     public int getUserId() { return userId; }
     public void setUserId(int userId) { this.userId = userId; }
 
-    public int getAmount() { return amount; }
-    public void setAmount(int amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
 
     public String getCurrency() { return currency; }
     public void setCurrency(String currency) { this.currency = currency; }
